Extract base class in order to share createAffectedTables and getAffectedTables methods
