+++ lib/cartodb/models/mapconfig/provider/map-store-provider.js
@@ -1,5 +1,5 @@
+const BaseMapConfigProvider = require('./base-mapconfig-adapter');
 const dot = require('dot');
-const QueryTables = require('cartodb-query-tables');
 
 // Configure bases for cache keys suitable for string interpolation
 const baseKey   = '{{=it.dbname}}:{{=it.token}}';
@@ -8,7 +8,7 @@ const rendererKey = baseKey + ':{{=it.dbuser}}:{{=it.format}}:{{=it.layer}}:{{=i
 const baseKeyTpl = dot.template(baseKey);
 const rendererKeyTpl = dot.template(rendererKey);
 
-module.exports = class MapStoreMapConfigProvider {
+module.exports = class MapStoreMapConfigProvider extends BaseMapConfigProvider {
     /**
      * @param {MapStore} mapStore
      * @param {String} user
@@ -18,15 +18,16 @@ module.exports = class MapStoreMapConfigProvider {
      * @type {MapStoreMapConfigProvider}
      */
     constructor (mapStore, user, userLimitsBackend, pgConnection, affectedTablesCache, params) {
+        super();
         this.mapStore = mapStore;
         this.user = user;
         this.userLimitsBackend = userLimitsBackend;
         this.pgConnection = pgConnection;
         this.affectedTablesCache = affectedTablesCache;
+        this.params = params;
         this.token = params.token;
         this.cacheBuster = params.cache_buster || 0;
         this.mapConfig = null;
-        this.params = params;
         this.context = null;
     }
 
@@ -81,67 +82,5 @@ module.exports = class MapStoreMapConfigProvider {
         }, this.params);
 
         return (base) ? baseKeyTpl(tplValues) : rendererKeyTpl(tplValues);
-    }
-
-    createAffectedTables (callback) {
-        this.getMapConfig((err, mapConfig) => {
-            if (err) {
-                return callback(err);
-            }
-
-            const { dbname } = this.params;
-            const token = mapConfig.id();
-
-            const queries = [];
-
-            mapConfig.getLayers().forEach(layer => {
-                queries.push(layer.options.sql);
-                if (layer.options.affected_tables) {
-                    layer.options.affected_tables.map(table => {
-                        queries.push(`SELECT * FROM ${table} LIMIT 0`);
-                    });
-                }
-            });
-
-            const sql = queries.length ? queries.join(';') : null;
-
-            if (!sql) {
-                return callback();
-            }
-
-            this.pgConnection.getConnection(this.user, (err, connection) => {
-                if (err) {
-                    return callback(err);
-                }
-
-                QueryTables.getAffectedTablesFromQuery(connection, sql, (err, affectedTables) => {
-                    if (err) {
-                        return callback(err);
-                    }
-
-                    this.affectedTablesCache.set(dbname, token, affectedTables);
-
-                    callback(err, affectedTables);
-                });
-            });
-        });
-    }
-
-    getAffectedTables (callback) {
-        this.getMapConfig((err, mapConfig) => {
-            if (err) {
-                return callback(err);
-            }
-
-            const { dbname } = this.params;
-            const token = mapConfig.id();
-
-            if (this.affectedTablesCache.hasAffectedTables(dbname, token)) {
-                const affectedTables = this.affectedTablesCache.get(dbname, token);
-                return callback(null, affectedTables);
-            }
-
-            return this.createAffectedTables(callback);
-        });
     }
 };
